fix: update curl when inputs are changed
fix: update curl when inputs are changed

fix: update curl when request body is changed

fix(update-curl): remove log
