+++ src/templates/server-template.js
@@ -1,6 +1,7 @@
 import { html } from 'lit';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js'; // eslint-disable-line import/extensions
 import { marked } from 'marked';
+import updateCurl from '~/utils/update-curl';
 
 export function setApiServer(serverUrl) {
   const serverObj = this.resolvedSpec?.servers.find((s) => s.url === serverUrl);
@@ -27,6 +28,8 @@ function onApiServerVarChange(e, serverObj) {
     tempUrl = tempUrl.replace(regex, v.value);
   });
   serverObj.computedUrl = tempUrl;
+  updateCurl.call(this, e.target ? e.target : e);
+
   this.requestUpdate();
 }
